Fix so that PLUCodedProducts use their mass correctly
Added productsPricedByMass hashmap, which is updated when PLUCodedItems
are added and removed. This allows us to keep track of the mass of items
which are added and do not have an expected mass in the product
database. This is then referenced when getExpectedMass and getTotalPrice
is called so they now account for PLUCodedProducts.